feat(gamification): decompose achievements screen into modular widgets (G4)
- Extracted AchievementsHeader, ChallengesSection, BadgesSection with BadgeCard & detail sheet
- Slimmed AchievementsScreen to use new widgets (<300 LOC)
- Updated tech plan doc to mark G4 completed
- Ensured analyzer clean and tests pass
